clean restructure; still more work to do
