+++ README.md
@@ -4,11 +4,6 @@ Using Go's http library to serve static files and assets
 
 To clear the local database and run in debug mode, run the following command:
 
-```bash
-go build -o out && ./out --debug
-```
-
-
 ## Whats included
 - HTTP using the STD
 - Simple storage using a `dataabse.json` file
@@ -17,12 +12,9 @@ go build -o out && ./out --debug
 
 
 ## TODO:
+- Add Service and Repository layers
+- Add Tests for services, handlers
 - Add sqlite3 database
 
-### Docker
-
-```bash
-GOOS=linux GOARCH=amd64 go build -o chirpy
-docker build . -t chirpy:latest
-docker run -p 8080:8080 chirpy:latest
-```
+### Running
+See `Makefile` for commands. Can be ran with `make run` or `make docker-run`
