+++ cmd/chirpy/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/ptenteromano/chirpy/internal/server"
+)
+
+func main() {
+	server.Start()
+}
