+++ internal/handlers/constants.go
@@ -1,0 +1,3 @@
+package handlers
+
+const MAX_CHIRP_LENGTH = 140
