+++ internal/handlers/health.go
@@ -1,0 +1,9 @@
+package handlers
+
+import "net/http"
+
+func HealthStatus(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write([]byte("OK"))
+}
